Moved Patterns namespace in its own file.
